+++ server/context_processors.py
@@ -14,3 +14,10 @@ def google_analytics(request):
         }
     else:
         return {"google_analytics": ""}
+
+
+def site_url(request):
+    """
+    Returns site URL.
+    """
+    return {"SITE_URL": settings.SITE_URL}