+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Collect email addresses from github logins
 - Allow users to drop files onto the homepage to create a new notebook with
   those files attached (#2133)
+- Preserve DOM changes in non-markdown/css cells while editing report
 
 # 0.13.0 (2019-10-01)